move constraints to its own package
