+++ contribs/drt/src/test/java/org/matsim/contrib/drt/run/DrtConfigGroupTest.java
@@ -2,8 +2,8 @@ package org.matsim.contrib.drt.run;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.matsim.contrib.drt.optimizer.DrtOptimizationConstraintsParams;
-import org.matsim.contrib.drt.optimizer.DrtOptimizationConstraintsSet;
+import org.matsim.contrib.drt.optimizer.constraints.DrtOptimizationConstraintsParams;
+import org.matsim.contrib.drt.optimizer.constraints.DrtOptimizationConstraintsSet;
 import org.matsim.core.config.Config;
 import org.matsim.core.config.ConfigUtils;
